Backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
